src/application/rules_api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-cloud-firestore==2.14.0
python-dotenv==1.0.1
pydantic==2.6.1 